+++ apiserver/service/src/main/java/com/zipline/service/survey/dto/request/SurveyCreateRequestDTO.java
@@ -1,4 +1,4 @@
-package com.zipline.dto.survey;
+package com.zipline.service.survey.dto.request;
 
 import java.util.ArrayList;
 import java.util.List;
