+++ apiserver/service/src/main/java/com/zipline/service/survey/dto/response/SurveyResponseDTO.java
@@ -1,4 +1,4 @@
-package com.zipline.dto.survey;
+package com.zipline.service.survey.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -6,6 +6,7 @@ import java.util.stream.Collectors;
 
 import com.zipline.entity.survey.Question;
 import com.zipline.entity.survey.Survey;
+
 import lombok.Getter;
 
 @Getter
